Updated term.

+++ demo.py
@@ -11,7 +11,7 @@ if len(sys.argv) == 1 :
     username = sys.stdin.readline().strip()
     print "Enter your WebAdvisor password:"
     pw = getpass.getpass().strip()
-    wave = WebAdvisor(username, pw, 's16')
+    wave = WebAdvisor(username, pw, 'f16')
     rosters = wave.get_rosters()
     print "Writing rosters.json"
     f = open('rosters.json', 'w')
